+++ packages/strapi-plugin-content-manager/admin/src/components/CustomTable/ActionCollapse/Wrapper.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+const Wrapper = styled.tr`
+  width: 100%;
+  height: 36px;
+  background: #f7f8f8;
+
+  td {
+    height: 36px;
+    line-height: 36px;
+    font-size: 1.3rem;
+    font-weight: 400;
+    color: #333740;
+    text-align: left;
+    border-collapse: collapse;
+    border-top: 1px solid #f1f1f2 !important;
+  }
+`;
+
+export default Wrapper;
