+++ packages/strapi-plugin-content-manager/admin/src/components/CustomTable/Row/ActionContainer.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+const ActionContainer = styled.td`
+  text-align: right;
+
+  i,
+  svg {
+    margin-left: 15px;
+    font-size: 1rem;
+
+    color: #333740;
+
+    &:first-of-type {
+      margin-left: 0px;
+    }
+  }
+`;
+
+export default ActionContainer;
