+++ packages/strapi-plugin-content-manager/admin/src/components/Truncate/index.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+const Truncate = styled.div``;
+
+export default Truncate;
