+++ packages/strapi-plugin-content-manager/admin/src/components/Truncated/index.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const Truncated = styled.p`
+  overflow-x: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  margin-bottom: 0;
+`;
+
+export default Truncated;
